Show error when something goes bad while deleting food

+++ src/foundation/food/RemoveFood.ts
@@ -5,17 +5,24 @@ import axios from "axios";
 export default class FindFood {
   private myCreatedFoodsRepository = new MyCreatedFoodsRepository();
 
-  async execute(food: FoodOffer | string): Promise<void> {
+  async execute(food: FoodOffer | string): Promise<boolean> {
     const effectiveFood =
       typeof food === "string"
         ? this.myCreatedFoodsRepository.find(food)
         : food;
 
     if (effectiveFood) {
-      await axios.delete(
+      const response = await axios.delete(
         `https://europe-west1-yotecocino-d6292.cloudfunctions.net/deleteOffer?id=${effectiveFood?.id}`
       );
-      this.myCreatedFoodsRepository.delete(effectiveFood);
+
+      if (response.status === 200) {
+        this.myCreatedFoodsRepository.delete(effectiveFood);
+      }
+      
+      return response.status === 200;
+    } else {
+      return false;
     }
   }
 }
